Improve optimized public administration showcase
